1.修复发送上级平台日期年份的数据类型 2.去除子模块改用nuget方式引用809库

+++ src/JT809.DotNetty.Simples/Inferior/JT809.Inferior.Client/JT809InferiorService.cs
@@ -51,7 +51,7 @@ namespace JT809.Inferior.Client
                             {
                                 Day = (byte)(DateTime.Now.Day),
                                 Month = (byte)(DateTime.Now.Month),
-                                Year = (byte)(DateTime.Now.Year),
+                                Year = (ushort)(DateTime.Now.Year),
                                 Hour = (byte)(DateTime.Now.Hour),
                                 Minute = (byte)(DateTime.Now.Minute),
                                 Second = (byte)(DateTime.Now.Second),
